Refactor language dialects into static backend layers
