+++ tilelang/env.py
@@ -394,7 +394,6 @@ class Environment:
     # Compilation defaults (for jit, autotune, compile)
     # These allow overriding default compilation parameters via environment variables
     TILELANG_DEFAULT_TARGET = EnvVar("TILELANG_DEFAULT_TARGET", "auto")
-    TILELANG_DEFAULT_DIALECT = EnvVar("TILELANG_DEFAULT_DIALECT", "core")
     TILELANG_DEFAULT_EXECUTION_BACKEND = EnvVar("TILELANG_EXECUTION_BACKEND", "auto")
     TILELANG_DEFAULT_VERBOSE = EnvVar("TILELANG_VERBOSE", "0")
