+++ tilelang/language/tir/ir.py
@@ -310,12 +310,6 @@ mma_fill = _dtype_forward(_tir_op.mma_fill)
 vectorlow = _dtype_forward(_tir_op.vectorlow)
 vectorhigh = _dtype_forward(_tir_op.vectorhigh)
 vectorcombine = _dtype_forward(_tir_op.vectorcombine)
-tvm_mfma = _dtype_forward(_tir_op.tvm_mfma)
-tvm_mfma_store = _dtype_forward(_tir_op.tvm_mfma_store)
-tvm_rdna_wmma = _dtype_forward(_tir_op.tvm_rdna_wmma)
-tvm_rdna_wmma_store = _dtype_forward(_tir_op.tvm_rdna_wmma_store)
-
-
 # ---------------------------------------------------------------------------
 # Cast with optional CUDA/PTX rounding hints.
 #
