+++ tilelang/webgpu/language/__init__.py
@@ -1,11 +1,11 @@
-"""WebGPU language dialect overlay."""
+"""WebGPU language dialect: common TileLang plus WebGPU extensions."""
 
 from __future__ import annotations
 
-from tilelang.language.dialect import export_core_language_symbols
-
-export_core_language_symbols(globals())
-del export_core_language_symbols
+from tilelang.language.common import *  # noqa: F401,F403
+from tilelang.language.common import __all__ as _COMMON_ALL
 
 __tilelang_dialect__ = "webgpu"
-__all__ = sorted(name for name in globals() if not (name.startswith("__") and name.endswith("__")))
+__all__ = tuple(_COMMON_ALL)
+
+del _COMMON_ALL
